requirements.txt
PySide6
keyboard
Pillow